+++ lib/api/resource.go
@@ -1,4 +1,4 @@
-package sebak
+package api
 
 import (
 	"github.com/nvellon/hal"
@@ -7,26 +7,34 @@ import (
 
 type APIResource interface {
 	LinkSelf() string
-	Resource(selfUrl string) *hal.Resource
+	Resource() *hal.Resource
 	GetMap() hal.Entry
+	Serialize() ([]byte, error)
 }
 
-type APIResourceList []APIResource
+type APIResourceList struct {
+	Resources []APIResource
+	SelfLink  string
+}
 
-func (al APIResourceList) Resource(selfUrl string) *hal.Resource {
-	rl := hal.NewResource(struct{}{}, selfUrl)
-	for _, apiResource := range al {
-		r := apiResource.Resource(apiResource.LinkSelf())
+func (al APIResourceList) Resource() *hal.Resource {
+	rl := hal.NewResource(struct{}{}, al.LinkSelf())
+	for _, apiResource := range al.Resources {
+		r := apiResource.Resource()
 		rl.Embed("records", r)
 	}
-	rl.AddLink("prev", hal.NewLink(selfUrl)) //TODO: set prev/next url
-	rl.AddLink("next", hal.NewLink(selfUrl))
+	rl.AddLink("prev", hal.NewLink(al.LinkSelf())) //TODO: set prev/next url
+	rl.AddLink("next", hal.NewLink(al.LinkSelf()))
 
 	return rl
 }
 
+func (al APIResourceList) Serialize() (encoded []byte, err error) {
+	return al.Resource().MarshalJSON()
+}
+
 func (al APIResourceList) LinkSelf() string {
-	return ""
+	return al.SelfLink
 }
 func (al APIResourceList) GetMap() hal.Entry {
 	return hal.Entry{}
@@ -53,8 +61,8 @@ func (aa APIResourceAccount) GetMap() hal.Entry {
 	}
 }
 
-func (aa APIResourceAccount) Resource(selfUrl string) *hal.Resource {
-	r := hal.NewResource(aa, selfUrl)
+func (aa APIResourceAccount) Resource() *hal.Resource {
+	r := hal.NewResource(aa, aa.LinkSelf())
 	r.AddLink("transactions", hal.NewLink(strings.Replace(UrlAccounts, "{id}", aa.accountId, -1)+"/transactions{?cursor,limit,order}", hal.LinkAttr{"templated": true}))
 	r.AddLink("operations", hal.NewLink(strings.Replace(UrlAccounts, "{id}", aa.accountId, -1)+"/operations{?cursor,limit,order}", hal.LinkAttr{"templated": true}))
 	return r
@@ -62,6 +70,10 @@ func (aa APIResourceAccount) Resource(selfUrl string) *hal.Resource {
 
 func (aa APIResourceAccount) LinkSelf() string {
 	return strings.Replace(UrlAccounts, "{id}", aa.accountId, -1)
+}
+
+func (aa APIResourceAccount) Serialize() (encoded []byte, err error) {
+	return aa.Resource().MarshalJSON()
 }
 
 type APIResourceTransaction struct {
@@ -89,9 +101,9 @@ func (at APIResourceTransaction) GetMap() hal.Entry {
 		"operation_count":   len(at.operations),
 	}
 }
-func (at APIResourceTransaction) Resource(selfUrl string) *hal.Resource {
+func (at APIResourceTransaction) Resource() *hal.Resource {
 
-	r := hal.NewResource(at, selfUrl)
+	r := hal.NewResource(at, at.LinkSelf())
 	r.AddLink("accounts", hal.NewLink(strings.Replace(UrlAccounts, "{id}", at.source, -1)))
 	r.AddLink("operations", hal.NewLink(strings.Replace(UrlTransactions, "{id}", at.hash, -1)+"/operations{?cursor,limit,order}", hal.LinkAttr{"templated": true}))
 	return r
@@ -99,6 +111,10 @@ func (at APIResourceTransaction) Resource(selfUrl string) *hal.Resource {
 
 func (at APIResourceTransaction) LinkSelf() string {
 	return strings.Replace(UrlTransactions, "{id}", at.hash, -1)
+}
+
+func (at APIResourceTransaction) Serialize() (encoded []byte, err error) {
+	return at.Resource().MarshalJSON()
 }
 
 type APIResourceOperation struct {
@@ -121,13 +137,17 @@ func (ao APIResourceOperation) GetMap() hal.Entry {
 	}
 }
 
-func (ao APIResourceOperation) Resource(selfUrl string) *hal.Resource {
+func (ao APIResourceOperation) Resource() *hal.Resource {
 
-	r := hal.NewResource(ao, selfUrl)
+	r := hal.NewResource(ao, ao.LinkSelf())
 	r.AddNewLink("transactions", strings.Replace(UrlTransactions, "{id}", ao.txHash, -1))
 	return r
 }
 
 func (ao APIResourceOperation) LinkSelf() string {
 	return strings.Replace(UrlOperations, "{id}", ao.hash, -1)
+}
+
+func (ao APIResourceOperation) Serialize() (encoded []byte, err error) {
+	return ao.Resource().MarshalJSON()
 }
